refactor: extract view from WorldManager
