+++ latebit/core/GameManager.test.cpp
@@ -15,7 +15,7 @@ auto main() -> int {
 
     Box wanted(Vector(0, 0), DM::WINDOW_WIDTH, DM::WINDOW_HEIGHT);
     assertEq("boundary is set up", WM.getBoundary(), wanted);
-    assertEq("view is set up", WM.getView(), wanted);
+    assertEq("view is set up", WM.getView().getView(), wanted);
     assertEq("is not over", GM.getGameOver(), false);
     assertEq("is not paused", GM.isPaused(), false);
