+++ latebit/core/world/WorldManager.cpp
@@ -8,21 +8,10 @@
 #include "latebit/core/objects/ObjectListIterator.h"
 #include "latebit/core/utils/utils.h"
 #include "latebit/utils/Logger.h"
-#include "latebit/utils/Math.h"
 
 #define WM lb::WorldManager::getInstance()
 
 namespace lb {
-
-auto WorldManager::worldToView(Vector worldPosition) -> Vector {
-  auto viewOrigin = WM.getView().getCorner();
-  return worldPosition - viewOrigin;
-}
-
-auto WorldManager::viewToWorld(Vector viewPosition) -> Vector {
-  auto viewOrigin = WM.getView().getCorner();
-  return viewPosition + viewOrigin;
-}
 
 WorldManager::WorldManager() : Manager("WorldManager") {
   Log.debug("WorldManager::WorldManager(): Created WorldManager");
@@ -168,12 +157,12 @@ void WorldManager::moveAndCheckBounds(Object* o, Vector position) {
     o->eventHandler(&event);
   }
 
-  if (this->viewFollowing == o) {
-    auto viewDeadZone = this->getViewDeadZone();
+  if (this->view.getViewFollowing() == o) {
+    auto viewDeadZone = this->view.getViewDeadZone();
 
     // Move the view if the object is outside of the dead zone
     if (!contains(viewDeadZone, final)) {
-      this->setViewPosition(position);
+      this->view.setViewPosition(position);
     }
   }
 }
@@ -225,7 +214,7 @@ void WorldManager::draw() {
 
     for (iterator.first(); !iterator.isDone(); iterator.next()) {
       auto object = iterator.currentObject();
-      if (object != nullptr && intersects(object->getWorldBox(), this->view)) {
+      if (object != nullptr && intersects(object->getWorldBox(), this->view.getView())) {
         object->draw();
       };
     }
@@ -235,63 +224,6 @@ void WorldManager::draw() {
 void WorldManager::setBoundary(Box b) { this->boundary = b; }
 auto WorldManager::getBoundary() const -> Box { return this->boundary; }
 
-void WorldManager::setView(Box v) { this->view = v; }
-auto WorldManager::getView() const -> Box { return this->view; }
-
-auto WorldManager::setViewPosition(Vector where) -> void {
-  auto viewHeight = this->view.getHeight();
-  auto viewWidth = this->view.getWidth();
-  auto worldHeight = this->boundary.getHeight();
-  auto worldWidth = this->boundary.getWidth();
-  auto worldOrigin = this->boundary.getCorner();
-  auto worldOriginX = worldOrigin.getX();
-  auto worldOriginY = worldOrigin.getY();
-
-  // easier to understand as the following translations
-  // viewCenter = corner + Vector(width / 2, height / 2)
-  // translate = where - viewCenter (note, `where` is the center of the view)
-  // newCorner = viewCorner + translate
-  auto newCorner = where - Vector(viewWidth / 2, viewHeight / 2);
-
-  newCorner.setX(clamp(newCorner.getX(), worldOriginX,
-                       worldOriginX + worldWidth - viewWidth));
-  newCorner.setY(clamp(newCorner.getY(), worldOriginY,
-                       worldOriginY + worldHeight - viewHeight));
-
-  this->view = Box(newCorner, viewWidth, viewHeight);
-}
-
-auto WorldManager::setViewFollowing(Object* o) -> int {
-  if (o == nullptr) {
-    this->viewFollowing = nullptr;
-    return 0;
-  }
-
-  auto updates = getAllObjects();
-  auto iterator = ObjectListIterator(&updates);
-  for (iterator.first(); !iterator.isDone(); iterator.next()) {
-    if (iterator.currentObject() == o) {
-      this->viewFollowing = o;
-      return 0;
-    }
-  }
-
-  Log.error(
-    "WorldManager::setViewFollowing(): Object %s to be followed was not found",
-    o->toString().c_str());
-  return -1;
-}
-
-void WorldManager::setViewDeadZone(Box d) {
-  if (!contains(this->view, d)) {
-    Log.error("WorldManager::setViewDeadZone(): Dead zone larger than view");
-    return;
-  }
-
-  this->viewDeadZone = d;
-}
-
-auto WorldManager::getViewDeadZone() const -> Box { return this->viewDeadZone; }
-
 auto WorldManager::getSceneGraph() -> SceneGraph& { return this->sceneGraph; }
+auto WorldManager::getView() -> View& { return this->view; }
 }  // namespace lb
